+++ ci.py
@@ -6,9 +6,7 @@ from gcip2 import BaseLinux, BaseTask, GitlabCiBuilderImpl
 from gcip2.pipeline_core import (
     ArtifactsReports,
     ArtifactsReportsCoverage,
-    Default,
     GlobalVariables,
-    Image,
     JobBuilderImpl,
     Stage,
     TriggerIncludeArtifact,
@@ -26,29 +24,6 @@ class Stages(str, Enum):
     JOBS = Stage.JOBS.value
     UNIT_TESTS = "unit-tests"
     PUBLISH = "publish"
-
-
-workflow = Workflow(
-    name="default",
-    auto_cancel=WorkflowAutoCancel(
-        on_job_failure=WorkflowAutoCancelOnJobFailure.NONE,
-        on_new_commit=WorkflowAutoCancelOnNewCommit.NONE,
-    ),
-    rules=[
-        WorkflowRule(
-            if_='$CI_PIPELINE_SOURCE == "merge_request_event"',
-            when=WorkflowWhen.ALWAYS,
-            auto_cancel=WorkflowAutoCancel(
-                on_new_commit=WorkflowAutoCancelOnNewCommit.INTERRUPTIBLE,
-            ),
-        ),
-        WorkflowRule(
-            if_="$CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH",
-            when=WorkflowWhen.ALWAYS,
-        ),
-        WorkflowRule(when=WorkflowWhen.NEVER),
-    ],
-)
 
 
 class RunUnitTests(JobBuilderImpl):
@@ -109,7 +84,7 @@ class GitlabCi(GitlabCiBuilderImpl):
                 self.job(TriggerPipeline)
                 .apply()
                 .with_name(f"{test_name}/trigger-pipeline")
-                .with_needs([build_pipeline_job.model.name])  # type: ignore
+                .with_needs([build_pipeline_job.model.name])
             )
             trigger_pipeline_job.model.trigger.include = [  # type: ignore
                 TriggerIncludeArtifact(
@@ -180,15 +155,30 @@ class GitlabCi(GitlabCiBuilderImpl):
         self.model.variables = {
             "PY_COLORS": GlobalVariables(value="1"),
             "FORCE_COLOR": GlobalVariables(value="1"),
+            # "FF_TIMESTAMPS": GlobalVariables(value="true"),
         }
 
-        self.with_workflow(workflow=workflow)
-        self.with_default(
-            Default(
-                tags=["static-k8s"],
-                image=Image(
-                    name="pfeiffermax/python-poetry:1.17.0-poetry2.2.1-python3.12.12-trixie",
+        self.with_workflow(
+            workflow=Workflow(
+                name="default",
+                auto_cancel=WorkflowAutoCancel(
+                    on_job_failure=WorkflowAutoCancelOnJobFailure.NONE,
+                    on_new_commit=WorkflowAutoCancelOnNewCommit.NONE,
                 ),
+                rules=[
+                    WorkflowRule(
+                        if_='$CI_PIPELINE_SOURCE == "merge_request_event"',
+                        when=WorkflowWhen.ALWAYS,
+                        auto_cancel=WorkflowAutoCancel(
+                            on_new_commit=WorkflowAutoCancelOnNewCommit.INTERRUPTIBLE,
+                        ),
+                    ),
+                    WorkflowRule(
+                        if_="$CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH",
+                        when=WorkflowWhen.ALWAYS,
+                    ),
+                    WorkflowRule(when=WorkflowWhen.NEVER),
+                ],
             )
         )
         return self