+++ gcip2/pipeline_core/pipeline.py
@@ -3,6 +3,7 @@ from typing import Self
 
 from gcip2.pipeline_core import (
     Default,
+    Image,
     Job,
     JobBuilderImpl,
     Pipeline,
@@ -43,8 +44,21 @@ class PipelineBuilderImpl(PipelineBuilder):
         return self
 
     def build(self: Self) -> Pipeline:
+        if not self.model.default:
+            self.model.default = Default()
+
+        if not self.model.default.image:
+            image_data = self._config.compose.images.get("base_python")
+            if not image_data:
+                raise ValueError(f"image_data not found in {self._config.compose.path.name}")
+            self.model.default.image = Image(name=image_data.image)
+
+        if not self.model.default.tags:
+            self.model.default.tags = ["static-k8s"]
+
         if not self.model.stages:
             self.model.stages = [Stage.JOBS]
+
         return self.model.model_copy(deep=True)
 
     @staticmethod