+++ pipelines/checkstyle/ci.py
@@ -3,8 +3,6 @@ from typing import Self
 
 from gcip2 import GitlabCiBuilderImpl, PipelineBuilderImpl
 from gcip2.pipeline_core import (
-    Default,
-    Image,
     JobBuilderImpl,
     Workflow,
     WorkflowAutoCancel,
@@ -32,19 +30,11 @@ class PreCommit(JobBuilderImpl):
         return self
 
 
-default = Default(
-    tags=["static-k8s"],
-    image=Image(
-        name="pfeiffermax/python-poetry:1.17.0-poetry2.2.1-python3.12.12-trixie",
-    ),
-)
-
-
 class Pipeline(PipelineBuilderImpl):
     def apply(self: Self) -> Self:
         self.model.stages = [Stages.pre_commit]
         self.add_jobs((self.job(PreCommit).apply(),))
-        return self.with_default(default)
+        return self
 
 
 class GitlabCi(GitlabCiBuilderImpl):
@@ -70,5 +60,4 @@ class GitlabCi(GitlabCiBuilderImpl):
                 ],
             )
         )
-        self.with_default(default)
         return self