+++ pipelines/initialization/ci.py
@@ -3,8 +3,6 @@ from typing import Self
 
 from gcip2 import GitlabCiBuilderImpl, PipelineBuilderImpl
 from gcip2.pipeline_core import (
-    Default,
-    Image,
     JobBuilderImpl,
     Trigger,
     TriggerForward,
@@ -62,14 +60,6 @@ class TriggerInitializationPipeline(TriggerPipeline):
         )
 
 
-default = Default(
-    tags=["static-k8s"],
-    image=Image(
-        name="pfeiffermax/python-poetry:1.17.0-poetry2.2.1-python3.12.12-trixie",
-    ),
-)
-
-
 class Pipeline(PipelineBuilderImpl):
     def apply(self: Self) -> Self:
         self.model.stages = [Stages.initialization]
@@ -81,7 +71,6 @@ class Pipeline(PipelineBuilderImpl):
             )
         )
 
-        self.with_default(default)
         return self
 
 
@@ -108,5 +97,4 @@ class GitlabCi(GitlabCiBuilderImpl):
                 ],
             )
         )
-        self.with_default(default)
         return self