+++ tests/integration/data/checkstyle/ci.py
@@ -26,7 +26,6 @@ class PreCommit(JobBuilderImpl):
     def apply(self: Self) -> Self:
         self.model.name = "pre-commit"
         self.model.script = [
-            "ls -la",
             "pre-commit run -av",
         ]
         self.with_stage(Stages.pre_commit)
@@ -45,15 +44,18 @@ workflow = Workflow(
             when=WorkflowWhen.ALWAYS,
         ),
         WorkflowRule(
-            if_="$CI_COMMIT_TAG =~ '/^v\\d+\\.\\d+\\.\\d+$/'",
-            when=WorkflowWhen.ALWAYS,
-        ),
-        WorkflowRule(
             if_="$CI_COMMIT_BRANCH == $CI_DEFAULT_BRANCH",
             when=WorkflowWhen.ALWAYS,
         ),
         WorkflowRule(when=WorkflowWhen.NEVER),
     ],
+)
+
+default = Default(
+    tags=["static-k8s"],
+    image=Image(
+        name="ghcr.io/astral-sh/uv:python3.12-bookworm",
+    ),
 )
 
 
@@ -63,14 +65,7 @@ class Pipeline(PipelineBuilderImpl):
 
         self.model.jobs.append(self.job(PreCommit).apply())
 
-        self.with_default(
-            Default(
-                tags=["static-k8s"],
-                image=Image(
-                    name="pfeiffermax/python-poetry:1.17.0-poetry2.2.1-python3.12.12-trixie",
-                ),
-            )
-        )
+        self.with_default(default)
         self.with_workflow(workflow=workflow)
         return self
 
@@ -79,12 +74,5 @@ class GitlabCi(GitlabCiBuilderImpl):
     def apply(self: Self) -> Self:
         super(GitlabCi, self).apply()
         self.with_workflow(workflow=workflow)
-        self.with_default(
-            Default(
-                tags=["static-k8s"],
-                image=Image(
-                    name="pfeiffermax/python-poetry:1.17.0-poetry2.2.1-python3.12.12-trixie",
-                ),
-            )
-        )
+        self.with_default(default)
         return self